Remove implicit lower-casing of 'TargetFramework'/'RuntimeIdentifier'
Not that we actually require lowercase but it was assumed in the design doc.

In case-sensitive file-systems the output paths will change according to the
casing of the above properties. So, to make the output paths consistent, we
made it lowercase before appending to the paths.

This is not needed now, since we want keep, what is specified but, warn
about the casing instead of an implicit conversion.