Promote publish path to project root
This change essentially enables the 'PublishDir' to be placed in the project root
instead of being under the 'OutputPath' (aka "bin" folder) as a default.

Most users using Publish will always output them in the project root for easy access.

They always do either of the following:

1. `dotnet publish -o publish`
2. `msbuild -t:Publish -p:PublishDir=publish`

Instead of making users repeat these everyday scenarios, we can make it simpler by moving
the default 'PublishDir' to the project root alongside other output paths like "bin" and "obj".

Note: As this is a breaking change, the docs needs to updated as well.